Fix a bug in image_loader.cpp, alpha should be 1.0 when channel number is less than 4. Add map_saver_test.
