+++ map_server/test/map_saver_test.test
@@ -1,0 +1,3 @@
+<launch>
+    <test test-name="map_saver_test" pkg="map_server" type="map_saver_test"/>
+</launch>
